final commit

+++ controllers/controllerOrder.js
@@ -3,7 +3,8 @@ const { Order, User, Car, } = require('../models')
 class Controller {
   static async getOrder(req, res, next) {
     try {
-      let data = await Order.findAll({
+      const { isAdmin } = req.user
+      let options = {
         include: [{ 
           model: User, 
           attributes: ['username']
@@ -12,7 +13,13 @@ class Controller {
           attributes: ['name', 'carType']
         }],
         order: [['createdAt','DESC']]
-      })
+      }
+
+      if (!isAdmin) {
+        options.where = { userId: req.user.id }
+      }
+
+      let data = await Order.findAll(options)
       res.status(200).json({ statusCode: 200, message:'Fetched all records' , data })
     } catch (error) {
       next(error)
@@ -21,10 +28,19 @@ class Controller {
 
   static async postOrder(req, res, next) {
     try {
-      const { id } = req.user
+      const { isAdmin, id: userId } = req.user
       const data = req.body
 
-      const dataOrder = { name:data.name, description:data.description, price:+data.price, imgUrl:data.imgUrl, authorId:id, categoryId:(+data.categoryId>0?+data.categoryId:'')}
+      const dataOrder = { 
+        pickUpLoc:data.pickUpLoc, 
+        dropOffLoc:data.dropOffLoc, 
+        pickUpDate:new Date(data.pickUpDate), 
+        dropOffDate:new Date(data.dropOffDate), 
+        pickUpTime:new Date(data.pickUpTime),
+        carId:data.carId,
+        userId:(isAdmin?null:userId),
+        adminId:(isAdmin?userId:null)
+      }
       const createdOrder = await Order.create(dataOrder)
       
       res.status(201).json({ statusCode: 201, message:'New record has been created', data:createdOrder })
@@ -36,23 +52,24 @@ class Controller {
 
   static async getOrderById(req, res, next) {
       try {
-          let { id } = req.params
+        const { isAdmin } = req.user
+        let { id } = req.params
 
-          let data = await Order.findByPk(id,{
-            include: [{ 
-              model: User, 
-              attributes: ['username']
-            }, { 
-              model: Car, 
-              attributes: ['name', 'carType']
-            }],
-          })
-          let isFound = data
-          if (isFound) {
-              res.status(200).json({ statusCode: 200, message: `Fetched record with given id: ${id}`, data })
-          } else {
-              throw {name: 'RecordNotFound'}
-          }
+        let data = await Order.findByPk(id,{
+          include: [{ 
+            model: User, 
+            attributes: ['username']
+          }, { 
+            model: Car, 
+            attributes: ['name', 'carType']
+          }],
+        })
+        let isFound = data
+        if (isFound) {
+            res.status(200).json({ statusCode: 200, message: `Fetched record with given id: ${id}`, data })
+        } else {
+            throw {name: 'RecordNotFound'}
+        }
 
       } catch (error) {
           next(error)
@@ -61,24 +78,28 @@ class Controller {
 
   static async deleteOrderById(req, res, next) {
       try {
-          let { id } = req.params
-          let data = await Order.findByPk(id)
-          let isFound = data
-          await Order.destroy({ where: { id } })
-          if (!isFound) {
-              throw {name: 'RecordNotFound'}
-          } else {
-              res.status(200).json({ statusCode: 200, message: `Record with given id {${id}} has been deleted`, data })
-          }
+        const { isAdmin } = req.user
+        if (!isAdmin) throw {name: 'Forbidden'}
+
+        let { id } = req.params
+        let data = await Order.findByPk(id)
+        let isFound = data
+        await Order.destroy({ where: { id } })
+        if (!isFound) {
+            throw {name: 'RecordNotFound'}
+        } else {
+            res.status(200).json({ statusCode: 200, message: `Record with given id {${id}} has been deleted`, data })
+        }
       } catch (error) {
-          next(error)
+        next(error)
       }
   }
 
   static async editOrderById(req, res, next) {
     try {
+      const { isAdmin, id: userId } = req.user
       const { id: OrderId } = req.params
-      const { id } = req.user
+      if (!isAdmin) throw {name: 'Forbidden'}
 
       const data = req.body
       const dataOrder = { 
@@ -88,12 +109,12 @@ class Controller {
         dropOffDate:new Date(data.dropOffDate), 
         pickUpTime:new Date(data.pickUpTime),
         carId:data.carId,
-        userId:id,
-        adminId:(data.adminId?+data.adminId:null)
+        userId:(isAdmin?null:userId),
+        adminId:(isAdmin?userId:null)
       }
       const createdOrder = await Order.update(dataOrder, { where:{id:OrderId}, returning:true })
 
-      res.status(201).json({ statusCode: 201, message:`Record with given id: ${OrderId} has been updated`, createdOrder })
+      res.status(201).json({ statusCode: 201, message:`Record with given id: ${OrderId} has been updated`, createdOrder:createdOrder[1][0] })
     } catch (error) {
       next(error)
     }
